microservices/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
bcrypt==3.2.0
Flask-CORS==3.0.10
redis==7.0.11
pika==1.0.0
requests